requirements.txt
# Framework Web
Flask

# Para carregar variáveis de ambiente (segredos)
python-dotenv

# SDK do Google Cloud (para Firestore RF-002 e App Engine RNF)
google-cloud-firestore
google-cloud-storage 

# SDK do Google Cloud (para Firestore RF-002 e App Engine RNF)
google-cloud-firestore
google-cloud-storage 

# Bibliotecas para Login com Google (OAuth 2.0)
Authlib